Move `get_log_filter` to `lib.rs`
