+++ rust/headless-client/src/lib.rs
@@ -21,7 +21,7 @@ use std::{
     path::PathBuf,
 };
 use tokio::sync::mpsc;
-use tracing_subscriber::{Layer as _, Registry, fmt, layer::SubscriberExt as _};
+use tracing_subscriber::{EnvFilter, Layer as _, Registry, fmt, layer::SubscriberExt as _};
 
 mod clear_logs;
 /// Generate a persistent device ID, stores it to disk, and reads it back.
@@ -135,7 +135,7 @@ impl Callbacks for CallbackHandler {
 
 /// Sets up logging for stdout only, with INFO level by default
 pub fn setup_stdout_logging() -> Result<FilterReloadHandle> {
-    let directives = ipc_service::get_log_filter().context("Can't read log filter")?;
+    let directives = get_log_filter().context("Can't read log filter")?;
     let (filter, reloader) = firezone_logging::try_filter(&directives)?;
     let layer = fmt::layer()
         .event_format(firezone_logging::Format::new())
@@ -144,6 +144,36 @@ pub fn setup_stdout_logging() -> Result<FilterReloadHandle> {
     firezone_logging::init(subscriber)?;
 
     Ok(reloader)
+}
+
+/// Reads the log filter for the IPC service or for debug commands
+///
+/// e.g. `info`
+///
+/// Reads from:
+/// 1. `RUST_LOG` env var
+/// 2. `known_dirs::ipc_log_filter()` file
+/// 3. Hard-coded default `SERVICE_RUST_LOG`
+///
+/// Errors if something is badly wrong, e.g. the directory for the config file
+/// can't be computed
+pub(crate) fn get_log_filter() -> Result<String> {
+    #[cfg(not(debug_assertions))]
+    const DEFAULT_LOG_FILTER: &str = "info";
+    #[cfg(debug_assertions)]
+    const DEFAULT_LOG_FILTER: &str = "debug";
+
+    if let Ok(filter) = std::env::var(EnvFilter::DEFAULT_ENV) {
+        return Ok(filter);
+    }
+
+    if let Ok(filter) =
+        std::fs::read_to_string(known_dirs::ipc_log_filter()?).map(|s| s.trim().to_string())
+    {
+        return Ok(filter);
+    }
+
+    Ok(DEFAULT_LOG_FILTER.to_string())
 }
 
 #[cfg(test)]
